pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shopify-incremental-sync"
version = "0.1.0"
description = "Repo-as-source-of-truth incremental Shopify sync scaffold"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = ["scripts", "scripts.lib"]
